+++ minfraud/webservice.py
@@ -105,9 +105,8 @@ class Client(object):
             uri,
             json=cleaned_request,
             auth=(self._user_id, self._license_key),
-            headers=
-            {'Accept': 'application/json',
-             'User-Agent': self._user_agent()},
+            headers={'Accept': 'application/json',
+                     'User-Agent': self._user_agent()},
             timeout=self._timeout)
         if response.status_code == 200:
             return self._handle_success(response, uri, model_class)