+++ HISTORY.rst
@@ -3,6 +3,16 @@
 History
 -------
 
+* Added support for new minFraud Insights outputs. These are:
+     * ``/credit_card/brand``
+     * ``/credit_card/type``
+     * ``/device/id``
+     * ``/email/is_free``
+     * ``/email/is_high_risk``
+* ``input`` on the ``Warning`` response model has been replaced with
+  ``input_pointer``. The latter is a JSON pointer to the input that
+  caused the warning.
+
 0.2.0 (2015-08-10)
 ++++++++++++++++++